Save specific string to input$cookie if unauthorized
This is needed to be able to run the app locally as well as on the server. When
running locally, we want to fall back to logging in via credentials in
.synapseConfig since the session token cookie is not available. However since
it's not available, its value is "" which means we can never make it past the
first if block. With this change, when the app is running on the server and the
user isn't logged in, the shiny input gets the value "unauthorized" and prompts
the user to log in, but when running locally it gets the empty string and can
log in using the .synapseConfig